requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0

fastapi==0.78.0
uvicorn[standard]
aiofiles
python-multipart
torch
openai-whisper